Solve day 6 part 2

+++ src/day6/Runner.java
@@ -6,17 +6,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Runner {
-    private final int time;
-    private final int distance;
+    private final long time;
+    private final long distance;
 
-    public Runner(int time, int distance) {
+    public Runner(long time, long distance) {
         this.time = time;
         this.distance = distance;
     }
 
     long getNbWinners() {
-        int startLosers = 0;
-        for (int i = 1; i < this.time; i++) {
+        long startLosers = 0;
+        for (long i = 1; i < this.time; i++) {
             if (i * (this.time - i) > this.distance) {
                 break;
             } else {
